requirements/test.txt
pytest>=3.0
pytest-django==3.1.2